pyproject.toml
[project]
name = "redmine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.13",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.18",
    "pydantic-settings>=2.9.1",
]
